refactor(cds): 抽取 fetchWithLockRetry 共享 helper(Bugbot e0f66dce)
Bugbot Review 2026-05-06 e0f66dce 指出 lock-aware retry 在
WorktreeService.fetchWithLockRetry 和 computeSelfStatusPayload 内联两处
重复。SSOT 抽到 cds/src/services/git-fetch-retry.ts,两处都改为薄 wrapper /
直接调用,future 调整退避或 lock 正则只动一处。

变更:
- 新增 cds/src/services/git-fetch-retry.ts —— 导出 fetchWithLockRetry(shell, cwd, branch, opts)
- WorktreeService.fetchWithLockRetry 简化为 thin wrapper(保留私有方法名兼容旧 call site)
- computeSelfStatusPayload(branches.ts) 内联 retry 删除,改调共享 helper

验证:
- pnpm tsc --noEmit 通过
- pnpm vitest run 1142/1142 全绿(module-load smoke 自动捕获新文件加载)
