+++ cds/src/routes/branches.ts
@@ -29,6 +29,7 @@ import { isSafeGitRef } from '../services/github-webhook-dispatcher.js';
 import { buildPreviewUrl } from '../services/comment-template.js';
 import { computePreviewSlug } from '../services/preview-slug.js';
 import { maskSecrets as maskSecretsText, shouldMask } from '../services/secret-masker.js';
+import { fetchWithLockRetry } from '../services/git-fetch-retry.js';
 
 // ── Self-status SSE 模块级状态 ────────────────────────────────────────
 // 为什么放模块级而不是闭包内:
@@ -125,30 +126,20 @@ async function computeSelfStatusPayload(
     fetchOk = false;
     fetchError = 'skipped (snapshot uses cached refs)';
   } else if (branchIsSafe) {
-    // ⚠ Bugbot Review 2026-05-06 930c5f98: broadcastSelfStatus 在 webhook 风暴
-    // 期间会和 deploy worker 的 git fetch 抢同一个 .git/refs/remotes/.../<branch>.lock,
-    // 单次 exec 直接退非 0,前端拿到 fetchOk=false + 计数空。WorktreeService.fetchWithLockRetry
-    // 已有同样 retry 逻辑;这里 inline 一份(模块级 free function 不便复用 class private)。
-    // maxAttempts=3 与 worktree 一致;退避 2s/4s 累计 ≤ 6s,在 SSE 推送可接受范围内。
+    // ⚠ Bugbot Review 2026-05-06 930c5f98 + e0f66dce: broadcastSelfStatus 在
+    // webhook 风暴期间会和 deploy worker 的 git fetch 抢同一个
+    // .git/refs/remotes/.../<branch>.lock。共享 fetchWithLockRetry(SSOT 在
+    // services/git-fetch-retry.ts),与 WorktreeService 同语义,改一处生效全局。
     try {
-      const maxAttempts = 3;
-      let lastResult: Awaited<ReturnType<typeof shell.exec>> | null = null;
-      for (let attempt = 1; attempt <= maxAttempts; attempt++) {
-        lastResult = await shell.exec(
-          `git fetch origin ${currentBranch}`,
-          { cwd: repoRoot, timeout: 5_000 },
-        );
-        if (lastResult.exitCode === 0) break;
-        const errOut = ((lastResult.stderr || '') + (lastResult.stdout || ''));
-        const isLockErr = /cannot lock ref|unable to create.*lock/i.test(errOut);
-        if (!isLockErr || attempt === maxAttempts) break;
-        await new Promise((r) => setTimeout(r, 2_000 * attempt));
-      }
-      if (!lastResult || lastResult.exitCode !== 0) {
+      const fetchResult = await fetchWithLockRetry(
+        shell,
+        repoRoot,
+        currentBranch,
+        { timeoutMs: 5_000 },
+      );
+      if (fetchResult.exitCode !== 0) {
         fetchOk = false;
-        fetchError = lastResult
-          ? (lastResult.stderr || lastResult.stdout || '').trim().slice(0, 500)
-          : 'fetch returned no result';
+        fetchError = (fetchResult.stderr || fetchResult.stdout || '').trim().slice(0, 500);
       }
     } catch (err) {
       fetchOk = false;
